CSCI-620/project-4-adityaanilraut/requirements.txt
Django==4.2.7
python-chess==1.999
gunicorn==21.2.0
channels==4.0.0
channels-redis==4.1.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
daphne==4.0.0
whitenoise==6.6.0
